# Uqse lazy iteration for better performance across entire package