--- a/.github/workflows/github_ci.yml
+++ b/.github/workflows/github_ci.yml
@@ -10,6 +10,8 @@ jobs:
       matrix:
         version:
           - '1.6'
+          - 'latest'
+          - 'nightly'
         os:
           - ubuntu-latest
           - macOS-latest