--- a/Project.toml
+++ b/Project.toml
@@ -14,7 +14,7 @@ Yields = "d7e99b2f-e7f3-4d9e-9f01-2338fc023ad3"
 ActuaryUtilities = "^0.6, ^1,^2"
 MortalityTables = "^0.12, ^1, ^2"
 Transducers = "^0.4"
-Yields = "^0.2,^0.3,^0.4, 0.5, 0.6, 0.7, 0.8, 0.9,0.10"
+Yields = "^0.2,^0.3,^0.4, 0.5, 0.6, 0.7, 0.8, 0.9,0.10, 0.11"
 julia = "^1.5"
 
 [extras]