Will require a major version bump.

# Before
```
julia> @benchmark premium_net($lc)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  57.708 μs …   9.280 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     60.917 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   71.587 μs ± 149.221 μs  ┊ GC (mean ± σ):  7.16% ± 4.72%

  ▄█                                                            
  ███▇▆▄▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▂▂▁▂▂▂▂▂ ▂
  57.7 μs         Histogram: frequency by time          151 μs <

 Memory estimate: 111.69 KiB, allocs estimate: 1233.

 ```

 # Before (after upstream PRs)
 ```
julia> @btime LifeContingencies.A($lc,10)
  17.833 μs (383 allocations: 10.84 KiB)

julia> @btime LifeContingencies.a($lc,10)
  12.459 μs (487 allocations: 13.47 KiB)
 ```

 # After

```

julia> @btime LifeContingencies.A($lc,10)
  184.771 ns (0 allocations: 0 bytes)

julia> @btime LifeContingencies.a($lc,10)
  395.109 ns (30 allocations: 640 bytes)
```

Todo:

- [ ] Draft some release notes / "what's changed"
- [ ] Identify and eliminate remaining allocations
  - [ ] `probability` is source of most(?) allocations
- [x] make `n` a positional/dispatch argument for `Insurance`
- [x] Annuity refactor
  - [x] make `n` a positional argument and dispatch on it's presence (eliminate `Union{Nothing,Int}`s)
- [ ] Update benchmarks (has resulted in some upstream PRs, e.g. https://github.com/JuliaActuary/MortalityTables.jl/pull/117)
- [x] Tests to pass (mostly related to Annuity Refactor)