@alecloudenback 2021-12-10
After restarting session, `@benchmark($lc)` returns to a version where there are a lot of allocations. If I *revise* the `cashflows` function, then the allocations go away in subsequent benchmarks. 

```
julia> @benchmark premium_net($lc)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  12.042 μs …  2.006 ms  ┊ GC (min … max): 0.00% … 98.50%
 Time  (median):     12.250 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   12.607 μs ± 19.943 μs  ┊ GC (mean ± σ):  1.57% ±  0.99%

    ▄▅▇█▇ ▆▄▃ ▂▂▁           ▁ ▂▃▄ ▄▄▃ ▂▂▁                     ▂
  ▃██████▁███▁███▁███▁▆▆▇▁▆▇█▁███▁███▁███▁█▇▇▁▇▇▆▁▄▅▆▁▇▅▆▁▇█▇ █
  12 μs        Histogram: log(frequency) by time      13.9 μs <

 Memory estimate: 5.75 KiB, allocs estimate: 276.

julia> @benchmark premium_net($lc) # after revising/saving/Revise.jl the `cashflows` function
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  9.583 μs …  26.042 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     9.709 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   9.808 μs ± 850.412 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

    ▁ ▆  █ ▇ ▅  ▂          ▁  ▁                               ▂
  ▄▁█▁█▁▁█▁█▁█▁▁█▁▇▁▅▁▁▅▁▇▁█▁▁█▁█▁▁▆▁▆▁▅▁▁▆▁▇▁▇▁▁█▁█▁▇▁▁▆▁▆▁▃ █
  9.58 μs      Histogram: log(frequency) by time      10.6 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
 ```

 Is it an issue of invalidations?